implement test scenarios for missing base version
