+++ x-pack/solutions/security/test/security_solution_api_integration/test_suites/detections_response/rules_management/prebuilt_rules/customization_enabled/customization/index.ts
@@ -8,7 +8,8 @@
 import type { FtrProviderContext } from '../../../../../../ftr_provider_context';
 
 export default ({ loadTestFile }: FtrProviderContext): void => {
-  loadTestFile(require.resolve('./calculate_is_customized'));
-  loadTestFile(require.resolve('./customize_prebuilt_rules'));
+  loadTestFile(require.resolve('./detect_customization_with_base_version'));
+  loadTestFile(require.resolve('./detect_customization_without_base_version'));
   loadTestFile(require.resolve('./customize_via_bulk_editing'));
+  loadTestFile(require.resolve('./unaffected_fields'));
 };
